Move all avatar code to avatar module
